requirements.txt
flask
python-binance